+++ alert-flutter/android/app/src/main/kotlin/io/alert/alert_flutter/MainActivity.kt
@@ -1,5 +1,0 @@
-package io.alert.alert_flutter
-
-import io.flutter.embedding.android.FlutterActivity
-
-class MainActivity: FlutterActivity()